requirements.txt
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.3.0
qdrant-client>=1.7.0
pymupdf4llm>=0.0.5
unstructured[pdf]>=0.11.0
langchain>=0.1.0
langchain-community>=0.1.0
streamlit>=1.28.0
requests>=2.31.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
langchain-text-splitters
huggingface_hub
plotly>=5.0.0